+++ src/app/home/business/business.module.scss
@@ -130,3 +130,11 @@
 
   background: #F8F8F8
 }
+
+.loadingAvatar {
+  border-radius: 50%;
+  margin: 15px;
+  flex-shrink: 0;
+  flex-grow: 0;
+  background-color: rgb(189, 189, 189);
+}
