+++ src/components/UI/LoadingView/loading.module.scss
@@ -37,4 +37,3 @@
   align-items: flex-end;
   color: rgb(77, 77, 77);
 }
-
